requirements.txt
annotated-types==0.6.0
anyio==4.2.0
bcrypt==4.1.2
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.108.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
psycopg2==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.24
starlette==0.32.0.post1
typing_extensions==4.9.0
uvicorn==0.25.0
